Changes the package name for jassimp.
