+++ GVRf/Framework/src/org/gearvrf/jassimp/AiBuiltInWrapperProvider.java
@@ -38,9 +38,42 @@ THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ---------------------------------------------------------------------------
  */
+package org.gearvrf.jassimp;
+
+import java.nio.ByteBuffer;
 
 /**
- * Java binding for the Open Asset Import Library.
+ * Wrapper provider using jassimp built in types.
  */
-package org.gearvrf.vendor.jassimp;
+public final class AiBuiltInWrapperProvider implements
+        AiWrapperProvider<AiVector, AiMatrix4f, AiColor, AiNode, AiQuaternion> {
 
+    @Override
+    public AiVector wrapVector3f(ByteBuffer buffer, int offset,
+            int numComponents) {
+
+        return new AiVector(buffer, offset, numComponents);
+    }
+
+    @Override
+    public AiMatrix4f wrapMatrix4f(float[] data) {
+        return new AiMatrix4f(data);
+    }
+
+    @Override
+    public AiColor wrapColor(ByteBuffer buffer, int offset) {
+        return new AiColor(buffer, offset);
+    }
+
+    @Override
+    public AiNode wrapSceneNode(Object parent, Object matrix,
+            int[] meshReferences, String name) {
+
+        return new AiNode((AiNode) parent, matrix, meshReferences, name);
+    }
+
+    @Override
+    public AiQuaternion wrapQuaternion(ByteBuffer buffer, int offset) {
+        return new AiQuaternion(buffer, offset);
+    }
+}
