+++ GVRf/Framework/src/org/gearvrf/jassimp/AiPrimitiveType.java
@@ -38,82 +38,40 @@ THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 ---------------------------------------------------------------------------
  */
-package org.gearvrf.vendor.jassimp;
+package org.gearvrf.jassimp;
 
 import java.util.Set;
 
 /**
- * Status flags for {@link AiScene}s.
+ * Enumerates the types of geometric primitives supported by Assimp.
+ * <p>
  */
-public enum AiSceneFlag {
+public enum AiPrimitiveType {
     /**
-     * Specifies that the scene data structure that was imported is not
-     * complete.
-     * <p>
-     * 
-     * This flag bypasses some internal validations and allows the import of
-     * animation skeletons, material libraries or camera animation paths using
-     * Assimp. Most applications won't support such data.
+     * A point primitive.
      */
-    INCOMPLETE(0x1),
+    POINT(0x1),
 
     /**
-     * This flag is set by the validation (
-     * {@link AiPostProcessSteps#VALIDATE_DATA_STRUCTURE
-     * VALIDATE_DATA_STRUCTURE}) postprocess-step if the validation is
-     * successful.
-     * <p>
-     * 
-     * In a validated scene you can be sure that any cross references in the
-     * data structure (e.g. vertex indices) are valid.
+     * A line primitive.
      */
-    VALIDATED(0x2),
+    LINE(0x2),
 
     /**
-     * * This flag is set by the validation (
-     * {@link AiPostProcessSteps#VALIDATE_DATA_STRUCTURE
-     * VALIDATE_DATA_STRUCTURE}) postprocess-step if the validation is
-     * successful but some issues have been found.
-     * <p>
-     * 
-     * This can for example mean that a texture that does not exist is
-     * referenced by a material or that the bone weights for a vertex don't sum
-     * to 1.0 ... . In most cases you should still be able to use the import.
-     * This flag could be useful for applications which don't capture Assimp's
-     * log output.
+     * A triangular primitive.
      */
-    VALIDATION_WARNING(0x4),
+    TRIANGLE(0x4),
 
     /**
-     * This flag is currently only set by the
-     * {@link jassimp.AiPostProcessSteps#JOIN_IDENTICAL_VERTICES
-     * JOIN_IDENTICAL_VERTICES}.
+     * A higher-level polygon with more than 3 edges.
      * <p>
      * 
-     * It indicates that the vertices of the output meshes aren't in the
-     * internal verbose format anymore. In the verbose format all vertices are
-     * unique, no vertex is ever referenced by more than one face.
+     * A triangle is a polygon, but polygon in this context means
+     * "all polygons that are not triangles". The "Triangulate"-Step is provided
+     * for your convenience, it splits all polygons in triangles (which are much
+     * easier to handle).
      */
-    NON_VERBOSE_FORMAT(0x8),
-
-    /**
-     * Denotes pure height-map terrain data.
-     * <p>
-     * 
-     * Pure terrains usually consist of quads, sometimes triangles, in a regular
-     * grid. The x,y coordinates of all vertex positions refer to the x,y
-     * coordinates on the terrain height map, the z-axis stores the elevation at
-     * a specific point.
-     * <p>
-     * 
-     * TER (Terragen) and HMP (3D Game Studio) are height map formats.
-     * <p>
-     * Assimp is probably not the best choice for loading *huge* terrains -
-     * fully triangulated data takes extremely much free store and should be
-     * avoided as long as possible (typically you'll do the triangulation when
-     * you actually need to render it).
-     */
-    TERRAIN(0x10);
+    POLYGON(0x8);
 
     /**
      * Utility method for converting from c/c++ based integer enums to java
@@ -128,9 +86,9 @@ public enum AiSceneFlag {
      * @param rawValue
      *            an integer based enum value (as defined by assimp)
      */
-    static void fromRawValue(Set<AiSceneFlag> set, int rawValue) {
+    static void fromRawValue(Set<AiPrimitiveType> set, int rawValue) {
 
-        for (AiSceneFlag type : AiSceneFlag.values()) {
+        for (AiPrimitiveType type : AiPrimitiveType.values()) {
             if ((type.m_rawValue & rawValue) != 0) {
                 set.add(type);
             }
@@ -143,7 +101,7 @@ public enum AiSceneFlag {
      * @param rawValue
      *            maps java enum to c/c++ integer enum values
      */
-    private AiSceneFlag(int rawValue) {
+    private AiPrimitiveType(int rawValue) {
         m_rawValue = rawValue;
     }
